load balancing seems to work

+++ src/main/java/csx55/overlay/wireformats/TaskDelivery.java
@@ -70,7 +70,7 @@ public class TaskDelivery implements Event {
         return this.numberOfNodes;
     }
 
-    public double getNumTasks() {
+    public int getNumTasks() {
         return this.numTasks;
     }
